refactor: extract BioMarker type from Jotai atom to dedicated types file
