+++ src/processors/enrich/inferData.ts
@@ -1,6 +1,6 @@
 import { getPhenotypicAge, getPhenotypicAge2 } from "./phenoAge";
 import { labels } from "../../data";
-import { BioMarker } from "../../atom/dataAtom";
+import { BioMarker } from "../../types/biomarker";
 
 type Recipe = [string, string[], (...args: any[]) => string, Partial<BioMarker[3]>?];
